Port fts_fuzzy_match to Lua
I found your JS module for fuzzy match while looking for a good fuzzy search similar to sublime text's for a simple project. I ported it to Lua for use in my project and figured I'd share it here so that others can use it if they are working in Lua. I kept everything pretty much identical to the JS version except for syntax differences. The module has been used in a personal project and works great so far with no errors, but no actual testing code was written for it.